requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
asyncssh==2.14.2
atpublic==4.1.0
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
billiard==4.2.0
bleach==6.1.0
blinker==1.8.2
boto3==1.34.129
botocore==1.34.129
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cianparser==1.0.4
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudscraper==1.2.71
colorama==0.4.6
comm==0.2.2
configobj==5.0.8
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
DateTime==5.5
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==0.22.1
dvc==3.51.2
dvc-data==3.15.1
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.20.0
dvc-task==0.4.0
entrypoints==0.4
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.3
Flask==3.0.3
Flask-Cors==4.0.1
Flask-HTTPAuth==4.8.0
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.53.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.0
funcy==2.0
gitdb==4.0.11
GitPython==3.1.43
grandalf==0.8
gto==1.7.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hydra-core==1.3.2
idna==3.7
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
isoduration==20.11.0
iterative-telemetry==0.0.8
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
kombu==5.3.7
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
multidict==6.0.5
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.2.1
notebook_shim==0.2.4
numpy==2.0.0
omegaconf==2.3.0
orjson==3.10.5
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==3.11.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydot==2.0.0
pygit2==1.15.0
PyGithub==2.3.0
Pygments==2.18.0
pygtrie==2.5.0
PyJWT==2.8.0
PyNaCl==1.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.1
scikit-learn==1.5.0
scipy==1.13.1
scmrepo==3.3.5
semver==3.0.2
Send2Trash==1.8.3
setuptools==70.1.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
sqltrie==0.11.0
stack-data==0.6.3
tabulate==0.9.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
tomlkit==0.12.5
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transliterate==1.10.2
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
vine==5.1.0
voluptuous==0.14.2
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
widgetsnbextension==4.0.11
wrapt==1.16.0
yarl==1.9.4
zc.lockfile==3.0.post1
zope.interface==6.4.post2
